+++ tests/ctl/core/test_api.py
@@ -5,7 +5,7 @@ from typing import Dict
 
 import pytest
 import requests
-from fideslang import model_list, parse
+from fideslang import DEFAULT_TAXONOMY, model_list, parse
 from pytest import MonkeyPatch
 from starlette.testclient import TestClient
 
@@ -13,6 +13,8 @@ from fidesctl.api.ctl.routes import health
 from fidesctl.api.ctl.routes.util import API_PREFIX
 from fidesctl.ctl.core import api as _api
 from fidesctl.ctl.core.config import FidesctlConfig
+
+TAXONOMY_ENDPOINTS = ["data_category", "data_subject", "data_use", "data_qualifier"]
 
 
 # Helper Functions
@@ -155,6 +157,153 @@ class TestCrud:
         )
         print(result.text)
         assert result.status_code == 200
+
+
+@pytest.mark.integration
+class TestDefaultTaxonomyCrud:
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_cannot_delete_default(
+        self, test_config: FidesctlConfig, endpoint: str
+    ) -> None:
+        resource = getattr(DEFAULT_TAXONOMY, endpoint)[0]
+
+        result = _api.delete(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            resource_id=resource.fides_key,
+            headers=test_config.user.request_headers,
+        )
+        assert result.status_code == 403
+
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_can_update_default(
+        self, test_config: FidesctlConfig, endpoint: str
+    ) -> None:
+        """Should be able to update as long as `is_default` is not changing"""
+        resource = getattr(DEFAULT_TAXONOMY, endpoint)[0]
+        json_resource = resource.json(exclude_none=True)
+
+        result = _api.update(
+            url=test_config.cli.server_url,
+            headers=test_config.user.request_headers,
+            resource_type=endpoint,
+            json_resource=json_resource,
+        )
+        assert result.status_code == 200
+
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_can_upsert_default(
+        self, test_config: FidesctlConfig, endpoint: str
+    ) -> None:
+        """Should be able to upsert as long as `is_default` is not changing"""
+        resources = [r.dict() for r in getattr(DEFAULT_TAXONOMY, endpoint)[0:2]]
+        result = _api.upsert(
+            url=test_config.cli.server_url,
+            headers=test_config.user.request_headers,
+            resource_type=endpoint,
+            resources=resources,
+        )
+        assert result.status_code == 200
+
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_cannot_create_default_taxonomy(
+        self, test_config: FidesctlConfig, resources_dict: Dict, endpoint: str
+    ) -> None:
+        manifest = resources_dict[endpoint]
+        manifest.is_default = True
+        result = _api.create(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            json_resource=manifest.json(exclude_none=True),
+            headers=test_config.user.request_headers,
+        )
+        assert result.status_code == 403
+
+        _api.delete(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            resource_id=manifest.fides_key,
+            headers=test_config.user.request_headers,
+        )
+
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_cannot_upsert_default_taxonomy(
+        self, test_config: FidesctlConfig, resources_dict: Dict, endpoint: str
+    ) -> None:
+        manifest = resources_dict[endpoint]
+        manifest.is_default = True
+        result = _api.upsert(
+            url=test_config.cli.server_url,
+            headers=test_config.user.request_headers,
+            resource_type=endpoint,
+            resources=[manifest.dict()],
+        )
+        assert result.status_code == 403
+
+        _api.delete(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            resource_id=manifest.fides_key,
+            headers=test_config.user.request_headers,
+        )
+
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_cannot_update_is_default(
+        self, test_config: FidesctlConfig, resources_dict: Dict, endpoint: str
+    ) -> None:
+        manifest = resources_dict[endpoint]
+        _api.create(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            json_resource=manifest.json(exclude_none=True),
+            headers=test_config.user.request_headers,
+        )
+
+        manifest.is_default = True
+        result = _api.update(
+            url=test_config.cli.server_url,
+            headers=test_config.user.request_headers,
+            resource_type=endpoint,
+            json_resource=manifest.json(exclude_none=True),
+        )
+        assert result.status_code == 403
+
+    @pytest.mark.parametrize("endpoint", TAXONOMY_ENDPOINTS)
+    def test_api_cannot_upsert_is_default(
+        self, test_config: FidesctlConfig, resources_dict: Dict, endpoint: str
+    ) -> None:
+        manifest = resources_dict[endpoint]
+        manifest.is_default = True
+        second_item = manifest.copy()
+        second_item.is_default = False
+
+        _api.create(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            json_resource=second_item.json(exclude_none=True),
+            headers=test_config.user.request_headers,
+        )
+
+        result = _api.upsert(
+            url=test_config.cli.server_url,
+            headers=test_config.user.request_headers,
+            resource_type=endpoint,
+            resources=[manifest.dict(), second_item.dict()],
+        )
+        assert result.status_code == 403
+
+        _api.delete(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            resource_id=manifest.fides_key,
+            headers=test_config.user.request_headers,
+        )
+        _api.delete(
+            url=test_config.cli.server_url,
+            resource_type=endpoint,
+            resource_id=second_item.fides_key,
+            headers=test_config.user.request_headers,
+        )
 
 
 # This test will fail if certain other tests run before it, due to a non-deterministic bug in the code